refactor(metor): Create useProducts for calling products
